# Refactor Agent OS Hooks Architecture for Performance and Modularity