… .agent-os/product/roadmap.md
… .agent-os/specs/2025-08-24-transparent-work-sessions-#75/spec.md
… .agent-os/specs/2025-08-24-transparent-work-sessions-#75/sub-specs/technical-spec.md
… .agent-os/specs/2025-08-24-transparent-work-sessions-#75/sub-specs/tests.md
… .agent-os/specs/2025-08-24-transparent-work-sessions-#75/tasks.md
… .agent-os/specs/2025-08-25-installation-test-suite-#76/spec.md
… .agent-os/specs/2025-08-25-installation-test-suite-#76/tasks.md
… CHANGELOG.md
… CLAUDE.md
… IMPLEMENTATION-PLAN.md
… PROJECT_INDEX.json
… claude-code/user/CLAUDE.md
… commands/hygiene-check.md
… docs/AUDIT_RESULTS.md
… docs/CLEANUP_TECH_DEBT.md
… docs/COMMAND-DISTINCTION.md
… docs/CONTEXT-TROUBLESHOOTING.md
… docs/IMPORTANT-NO-AOS-BACKGROUND.md
… hooks/MODULAR-ARCHITECTURE.md
… hooks/modules/posttool_handler.py
… hooks/modules/posttool_handler_optimized.py
… hooks/modules/pretool_handler.py
… hooks/modules/pretool_handler_optimized.py
… hooks/modules/task_handler.py
… hooks/modules/task_handler_optimized.py
… hooks/modules/userprompt_handler.py
… hooks/modules/userprompt_handler_optimized.py
… hooks/shared/cache_manager.py
… hooks/shared/git_utils.py
… hooks/tests/modular/benchmark_final.py